requirements.txt
fastapi
uvicorn
pydantic~=2.12.5
python-dotenv~=1.2.2
supabase~=2.28.0